fixed issue dragging splash dialog if downloading
Fixed issue that was causing an unhandled exception (stack overflow) if
the user dragged the splash while the engine was downloading a file.

Changed how web-service errors are logged to give priority to http
issues (if they occur)

Updated version to 14.1.1.0, meaning 2014, first semester, release 1.

If, for whatever reason, the engine cannot retrieve update info from the
server, it will assume that it is up-to-date.