adding rudimentary events
events will be raised when tasks / branches start and complete. Working
towards a way to update the user interface in javascript without having
to rely on a lot of separate ui xml tasks.  That is, we want to be able
to script a page to watch for events and update itself accordingly