adding redirect urls to embedded resources
If the engine sees a request for a file that matches the name of one of
its embedded html resource files, it will redirect that request to the
embedded file.  The intent here is to allow the engine to substitute its
internal resources for external resources, if it has the resources
already.  For example, pages that reference jquery against the external
cdn would fetch jquery from the internal resource rather than the
cdn....