adding ability control how browser handles hyperlinks
Link policies determine how the embedded browser opens hyperlinks from
its hosted content.

can now specify one of four link policies:

1) All -- all hyperlinks will be opened in the embedded browser
2) None -- all hyperlinks will be opened in external browser
3) Internal -- 'resource://' and 'archive://' hyperlinks will be opened
in the embedded browser. All other hyperlinks will be opened in an
external browser window. This is default value.
4) InternalAndSameHost -- 'resource://' and 'archive://' hyperlinks will
be opened in the embedded browser, and hyperlinks with the same host as
the original url will be opened in embedded browser. All other
hyperlinks will be opened in an external browser window